[#3/edit] create viewmodel for repo data
