+++ app/src/main/java/org/android/go/sopt/presentation/MainActivity.kt
@@ -1,21 +1,14 @@
 package org.android.go.sopt.presentation
 
-import android.annotation.SuppressLint
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
-import android.text.TextUtils.replace
-import android.view.View
-import androidx.activity.viewModels
-import androidx.core.view.isInvisible
-import androidx.databinding.adapters.TextViewBindingAdapter.setText
 import androidx.fragment.app.Fragment
 import org.android.go.sopt.R
 import org.android.go.sopt.data.User
-import org.android.go.sopt.databinding.ActivityJoinBinding
 import org.android.go.sopt.databinding.ActivityMainBinding
-import org.android.go.sopt.presentation.home.GalleryFragment
-import org.android.go.sopt.presentation.home.HomeFragment
-import org.android.go.sopt.presentation.home.SearchFragment
+import org.android.go.sopt.presentation.main.gallery.GalleryFragment
+import org.android.go.sopt.presentation.main.home.HomeFragment
+import org.android.go.sopt.presentation.main.search.SearchFragment
 
 class MainActivity : AppCompatActivity() {
     private var userData : User? = null
