+++ app/src/main/java/org/android/go/sopt/presentation/main/home/HomeAdapter.kt
@@ -1,10 +1,8 @@
-package org.android.go.sopt.presentation.home
+package org.android.go.sopt.presentation.main.home
 
 import android.content.Context
 import android.view.LayoutInflater
 import android.view.ViewGroup
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
 import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
@@ -17,16 +15,6 @@ import org.android.go.sopt.databinding.HomeItemRepoBinding
 class HomeAdapter(context: Context) : ListAdapter<HomeItem.Repo, RecyclerView.ViewHolder>(diffUtil) {
 
     private val inflater by lazy {LayoutInflater.from(context)}
-    private val item: List<Any> = listOf(
-        HomeItem.Repo(R.drawable.userpic, "git1","author1"),
-        HomeItem.Repo(R.drawable.userpic, "git2","author2"),
-        HomeItem.Repo(R.drawable.userpic, "git3","author3"),
-        HomeItem.Repo(R.drawable.userpic, "git4","author4"),
-        HomeItem.Repo(R.drawable.userpic, "git5","author5"),
-        HomeItem.Repo(R.drawable.userpic, "git6","author6"),
-        HomeItem.Repo(R.drawable.userpic, "git7","author7")
-
-    )
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): HomeViewHolder {
         return when (viewType) {
@@ -47,7 +35,7 @@ class HomeAdapter(context: Context) : ListAdapter<HomeItem.Repo, RecyclerView.Vi
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         when (holder) {
             is HomeViewHolder.RepoViewHolder ->
-                holder.bind(item[position-1] as HomeItem.Repo)
+                holder.bind(getItem(position) as HomeItem.Repo)
             else -> {}
         }
     }
@@ -57,9 +45,6 @@ class HomeAdapter(context: Context) : ListAdapter<HomeItem.Repo, RecyclerView.Vi
              0 -> HEADER_VIEW_TYPE
             else -> REPO_VIEW_TYPE
         }
-    }
-    override fun getItemCount(): Int {
-        return item.size + 1
     }
 
     companion object {
