+++ app/src/main/java/org/android/go/sopt/presentation/main/home/HomeFragment.kt
@@ -1,13 +1,13 @@
-package org.android.go.sopt.presentation.home
+package org.android.go.sopt.presentation.main.home
 
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.recyclerview.widget.RecyclerView
-import org.android.go.sopt.R
+import androidx.fragment.app.viewModels
 import org.android.go.sopt.databinding.FragmentHomeBinding
+import org.android.go.sopt.presentation.main.home.HomeAdapter
 import org.android.go.sopt.util.extension.scrollToTop
 
 class HomeFragment : Fragment() {
@@ -15,6 +15,7 @@ class HomeFragment : Fragment() {
     private var _binding: FragmentHomeBinding?= null
     private val binding: FragmentHomeBinding
         get() = requireNotNull(_binding)
+    private val viewmodel by viewModels<HomeViewModel>()
 
 
     override fun onCreateView(
@@ -23,17 +24,21 @@ class HomeFragment : Fragment() {
     ): View {
         _binding = FragmentHomeBinding.inflate(inflater,container,false)
         return binding.root
-        // Inflate the layout for this fragment
     }
 
     override fun onViewCreated(view: View, savedInstanceState: Bundle?) {
         super.onViewCreated(view, savedInstanceState)
         binding.rvHome.scrollToTop()
-        binding.rvHome.adapter = HomeAdapter(requireContext())
+        initAdapter(HomeAdapter(requireContext()))
     }
 
     override fun onDestroy() {
         super.onDestroy()
         _binding = null
+    }
+
+    private fun initAdapter(adapter: HomeAdapter){
+        adapter.submitList(viewmodel.fakeRepoList)
+        binding.rvHome.adapter = adapter
     }
 }
