+++ app/src/main/res/layout/activity_main.xml
@@ -16,7 +16,7 @@
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintBottom_toTopOf="@id/navi_main"
         android:tag="my_tag"
-        android:name="org.android.go.sopt.presentation.home.HomeFragment">
+        android:name="org.android.go.sopt.presentation.main.home.HomeFragment">
 
     </androidx.fragment.app.FragmentContainerView>
