+++ app/src/main/res/layout/home_item_header.xml
@@ -7,11 +7,13 @@
     <TextView
         android:id="@+id/item_recycler_header"
         android:layout_width="match_parent"
-        android:text="Repo List"
+        android:text="@string/home_header"
         android:layout_height="wrap_content"
         android:layout_marginVertical="10dp"
-        android:layout_marginHorizontal="50dp"
+        android:layout_marginHorizontal="10dp"
         android:gravity="start"
+        android:textSize="20sp"
+        android:textStyle="bold"
         app:layout_constraintTop_toTopOf="parent"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
